+++ src/main/java/org/truffleruby/core/kernel/TraceManager.java
@@ -65,7 +65,7 @@ public class TraceManager {
     }
 
     @TruffleBoundary
-    public void setTraceFunc(final DynamicObject traceFunc) {
+    public void setTraceFunc(DynamicObject traceFunc) {
         assert traceFunc == null || RubyGuards.isRubyProc(traceFunc);
 
         if (instruments != null) {